requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
setuptools==69.5.1
tk==0.1.0
wheel==0.43.0
